requirements.txt
h2o==3.44.0.2
pandas>=2.0.0
numpy>=1.24.0
shap>=0.46.0
google-generativeai>=0.7.0
streamlit>=1.35.0
scikit-learn
xgboost
joblib
matplotlib>=3.8

scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
shap>=0.48.0
streamlit>=1.32
joblib>=1.3
google-generativeai>=0.8.0
# Optional: only needed if you plan to run MOJO locally (and have Java 11)
# h2o==3.44.0.2